chore(npm): move examples
